支持float读写

+++ 05_模版/04_Json配置读取器/main.cpp
@@ -9,7 +9,7 @@ struct Student {
 
 	std::string name;
 	unsigned age;
-	double height;
+	float height;
 	bool canSwim;
 };
